+++ FreeRTOS-Plus/Source/FreeRTOS-Plus-TCP/FreeRTOS_Sockets.c
@@ -3419,12 +3419,12 @@ void vSocketWakeUpUser( FreeRTOS_Socket_t *pxSocket )
 				}
 				FreeRTOS_printf( ( "TCP %5d %-16lxip:%5d %d/%d %-13.13s %6lu %6u%s\n",
 					pxSocket->usLocalPort,		/* Local port on this machine */
-					pxSocket->u.xTCP.ulRemoteIP,	/* IP address of remote machine */
+					( unsigned long ) pxSocket->u.xTCP.ulRemoteIP,	/* IP address of remote machine */
 					pxSocket->u.xTCP.usRemotePort,	/* Port on remote machine */
 					pxSocket->u.xTCP.rxStream != NULL,
 					pxSocket->u.xTCP.txStream != NULL,
 					FreeRTOS_GetTCPStateName( pxSocket->u.xTCP.ucTCPState ),
-					(age > 999999 ? 999999 : age), /* Format 'age' for printing */
+					(age > 999999 ? 999999UL : age), /* Format 'age' for printing */
 					pxSocket->u.xTCP.usTimeout,
 					ucChildText ) );
 					/* Remove compiler warnings if FreeRTOS_debug_printf() is not defined. */
@@ -3443,9 +3443,9 @@ void vSocketWakeUpUser( FreeRTOS_Socket_t *pxSocket )
 			}
 
 			FreeRTOS_printf( ( "FreeRTOS_netstat: %lu sockets %lu < %lu < %d buffers free\n",
-				count,
-				uxGetMinimumFreeNetworkBuffers( ),
-				uxGetNumberOfFreeNetworkBuffers( ),
+				( unsigned long ) count,
+				( unsigned long ) uxGetMinimumFreeNetworkBuffers( ),
+				( unsigned long ) uxGetNumberOfFreeNetworkBuffers( ),
 				ipconfigNUM_NETWORK_BUFFER_DESCRIPTORS ) );
 		}
 	}
